refactor: restructure and use i18n
