+++ app/_layout.tsx
@@ -4,10 +4,25 @@ import * as SplashScreen from 'expo-splash-screen';
 import { useEffect } from 'react';
 import 'react-native-reanimated';
 import { PaperProvider } from 'react-native-paper';
-import Themes from '@/styles/themes';
+import { Themes } from '@/styles';
+import i18n from 'i18next';
+import Locales from '@/locales';
+import { initReactI18next } from 'react-i18next';
 
 // Prevent the splash screen from auto-hiding before asset loading is complete.
 SplashScreen.preventAutoHideAsync();
+i18n.use(initReactI18next).init({
+    lng: 'en',
+    fallbackLng: 'en',
+    compatibilityJSON: 'v3',
+    debug: true,
+    resources: {
+        ...Locales,
+    },
+    interpolation: {
+        escapeValue: false, // not needed for react as it escapes by default
+    },
+});
 
 export default function RootLayout() {
     const [loaded] = useFonts({
