+++ components/index.ts
@@ -1,0 +1,9 @@
+/**
+ * Components are single use, reusable, and composable building blocks.
+ * They should have a single responsibility and be as small as possible.
+ */
+
+import TabBar from './navigation/TabBar';
+import TabBarIcon from './navigation/TabBarIcon';
+
+export { TabBar, TabBarIcon };
