+++ components/navigation/TabBarIcon.tsx
@@ -4,6 +4,9 @@ import Ionicons from '@expo/vector-icons/Ionicons';
 import { type IconProps } from '@expo/vector-icons/build/createIconSet';
 import { type ComponentProps } from 'react';
 
-export function TabBarIcon({ style, ...rest }: IconProps<ComponentProps<typeof Ionicons>['name']>) {
+export default function TabBarIcon({
+    style,
+    ...rest
+}: IconProps<ComponentProps<typeof Ionicons>['name']>) {
     return <Ionicons size={28} style={[{ marginBottom: -3 }, style]} {...rest} />;
 }
