+++ styles/index.ts
@@ -1,0 +1,4 @@
+import Colors from './colors';
+import Themes from './themes';
+
+export { Colors, Themes };
